+++ internal/controllers/imageupdateautomation_controller.go
