+++ internal/controllers/imageupdateautomation_controller_test.go
