+++ internal/controllers/testdata/pathconfig/no/deploy.yaml
