+++ internal/controllers/testdata/pathconfig/yes/deploy.yaml
